requirements.txt
flask
reportlab
pillow
numpy
gunicorn
